refactor: refactor `get_ffa_feature_test_target`
Change `get_ffa_feature_test_target` to return a `size_t` instead of an
`unsigned int`, because `size_t` is the return type of operators like
`sizeof()`.

Change `get_ffa_feature_test_target` to require its argument to be
non-null (and assert that it is). This function is only used for getting
the array of features to test, so there is no use case where passing a
non-null pointer would make sense.

+++ spm/common/sp_tests/sp_test_ffa.c
@@ -82,8 +82,7 @@ static void ffa_features_test(void)
 {
 	const struct ffa_features_test *func_id_targets;
 	/* Get common features between tftf and cactus. */
-	unsigned int test_target_size =
-		get_ffa_feature_test_target(&func_id_targets);
+	size_t test_target_size = get_ffa_feature_test_target(&func_id_targets);
 	const struct ffa_features_test feature_id_targets[] = {
 		{"FFA_FEATURE_MEI", FFA_FEATURE_MEI, FFA_SUCCESS_SMC32, 0,
 			MAKE_FFA_VERSION(1, 1)},
